feat: move user store to store
