+++ package.json
@@ -1,69 +1,69 @@
 {
-	"name": "speak-georgian-app",
-	"private": true,
-	"version": "0.1.0",
-	"type": "module",
-	"scripts": {
-		"prepare": "husky",
-		"dev": "vite",
-		"build": "tsc && vite build",
-		"preview": "vite preview",
-		"test": "jest",
-		"test:watch": "jest --watch",
-		"check:types": "tsc",
-		"check:lint": "biome check --write ./src",
-		"check:deps": "knip"
-	},
-	"dependencies": {
-		"@hookform/resolvers": "^3.9.0",
-		"@lukemorales/query-key-factory": "^1.3.4",
-		"@nextui-org/react": "^2.4.6",
-		"@supabase/supabase-js": "^2.45.3",
-		"@tanstack/react-query": "^5.53.1",
-		"@tanstack/react-query-devtools": "^5.53.1",
-		"@tanstack/react-query-persist-client": "^5.53.1",
-		"@tanstack/react-router": "^1.52.3",
-		"framer-motion": "^11.3.31",
-		"i18next": "^23.14.0",
-		"i18next-browser-languagedetector": "^8.0.0",
-		"idb-keyval": "^6.2.1",
-		"lodash": "^4.17.21",
-		"lucide-react": "^0.309.0",
-		"react": "^18.3.1",
-		"react-dom": "^18.3.1",
-		"react-hook-form": "^7.53.0",
-		"react-hot-toast": "^2.4.1",
-		"react-i18next": "^14.1.3",
-		"zod": "^3.23.8",
-		"zustand": "^4.5.5"
-	},
-	"devDependencies": {
-		"@biomejs/biome": "1.8.2",
-		"@hookform/devtools": "^4.3.1",
-		"@tanstack/router-devtools": "^1.57.13",
-		"@tanstack/router-plugin": "^1.52.0",
-		"@testing-library/jest-dom": "^6.5.0",
-		"@testing-library/react": "^14.3.1",
-		"@total-typescript/ts-reset": "^0.5.1",
-		"@types/jest": "^29.5.12",
-		"@types/lodash": "^4.17.7",
-		"@types/react": "^18.3.5",
-		"@types/react-dom": "^18.3.0",
-		"@vitejs/plugin-react": "^4.3.1",
-		"autoprefixer": "^10.4.20",
-		"husky": "^9.1.5",
-		"jest": "^29.7.0",
-		"jest-environment-jsdom": "^29.7.0",
-		"knip": "^5.29.1",
-		"postcss": "^8.4.42",
-		"tailwind-merge": "^2.5.2",
-		"tailwind-scrollbar-hide": "^1.1.7",
-		"tailwindcss": "^3.4.10",
-		"tailwindcss-safe-area": "^0.4.1",
-		"ts-jest": "^29.2.5",
-		"typescript": "^5.5.4",
-		"vite": "^5.4.2",
-		"vite-plugin-svgr": "^4.2.0",
-		"vite-tsconfig-paths": "^4.3.2"
-	}
+  "name": "speak-georgian-app",
+  "private": true,
+  "version": "0.1.0",
+  "type": "module",
+  "scripts": {
+    "prepare": "husky",
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "check:types": "tsc",
+    "check:lint": "biome check --write ./src",
+    "check:deps": "knip"
+  },
+  "dependencies": {
+    "@hookform/resolvers": "^3.9.0",
+    "@lukemorales/query-key-factory": "^1.3.4",
+    "@nextui-org/react": "^2.4.8",
+    "@supabase/supabase-js": "^2.45.4",
+    "@tanstack/react-query": "^5.56.2",
+    "@tanstack/react-query-devtools": "^5.56.2",
+    "@tanstack/react-query-persist-client": "^5.56.2",
+    "@tanstack/react-router": "^1.57.15",
+    "framer-motion": "^11.5.4",
+    "i18next": "^23.15.1",
+    "i18next-browser-languagedetector": "^8.0.0",
+    "idb-keyval": "^6.2.1",
+    "lodash": "^4.17.21",
+    "lucide-react": "^0.309.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-hook-form": "^7.53.0",
+    "react-hot-toast": "^2.4.1",
+    "react-i18next": "^14.1.3",
+    "zod": "^3.23.8",
+    "zustand": "^4.5.5"
+  },
+  "devDependencies": {
+    "@biomejs/biome": "1.8.2",
+    "@hookform/devtools": "^4.3.1",
+    "@tanstack/router-devtools": "^1.57.13",
+    "@tanstack/router-plugin": "^1.52.0",
+    "@testing-library/jest-dom": "^6.5.0",
+    "@testing-library/react": "^14.3.1",
+    "@total-typescript/ts-reset": "^0.5.1",
+    "@types/jest": "^29.5.12",
+    "@types/lodash": "^4.17.7",
+    "@types/react": "^18.3.5",
+    "@types/react-dom": "^18.3.0",
+    "@vitejs/plugin-react": "^4.3.1",
+    "autoprefixer": "^10.4.20",
+    "husky": "^9.1.5",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "knip": "^5.30.2",
+    "postcss": "^8.4.47",
+    "tailwind-merge": "^2.5.2",
+    "tailwind-scrollbar-hide": "^1.1.7",
+    "tailwindcss": "^3.4.10",
+    "tailwindcss-safe-area": "^0.4.1",
+    "ts-jest": "^29.2.5",
+    "typescript": "^5.6.2",
+    "vite": "^5.4.5",
+    "vite-plugin-svgr": "^4.2.0",
+    "vite-tsconfig-paths": "^4.3.2"
+  }
 }
