+++ src/cache/category/useGetCategoryWords.ts
@@ -1,10 +1,10 @@
 import { useQuery } from '@tanstack/react-query';
-import { useUser } from 'src/auth/useUser';
+import { useUserStore } from 'src/store/user';
 import { getCategoryWords } from 'src/supabase/category/getCategoryWords';
 import { queryKeys } from '../keys';
 
 export const useGetCategoryWords = (categoryId?: number) => {
-	const user = useUser();
+	const { user } = useUserStore();
 
 	return useQuery({
 		queryKey: queryKeys.category.words(categoryId || null).queryKey,
