+++ src/cache/favorite/useGetFavoriteWords.ts
@@ -1,10 +1,10 @@
 import { useQuery } from '@tanstack/react-query';
-import { useUser } from 'src/auth/useUser';
+import { useUserStore } from 'src/store/user';
 import { getFavoriteWords } from 'src/supabase/favorite/getFavoriteWords';
 import { queryKeys } from '../keys';
 
 export const useGetFavoriteWords = () => {
-	const user = useUser();
+	const { user } = useUserStore();
 
 	return useQuery({
 		queryKey: queryKeys.favorite.words.queryKey,
